clean up search/save, move environment initialization to root
