+++ search2/requirements.txt
@@ -1,8 +1,9 @@
-maturin==0.14.13
-ray==2.3.0
--f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
-jax[cuda]==0.4.4
-pyarrow==11.0.0
+maturin==0.15.1
+ray==2.4.0
+jax==0.4.8
+jaxlib==0.4.7
+numpyro[cpu]==0.11.0
+pyarrow==12.0.0
 pystalk==0.7.0
 orjson==3.8.6
 ruamel.yaml==0.17.21
@@ -12,3 +13,5 @@ mgzip==0.2.1
 Cython==0.29.33
 patchelf==0.17.2
 aiohttp==3.8.4
+mypy==1.2.0
+pylint==2.17.4
